+++ src/core/data/modules/base.ts
@@ -103,7 +103,7 @@ export default abstract class Provider extends ParamsProvider implements IProvid
 		super();
 
 		const
-			id = this.getCacheKey(Object.select(opts, 'externalRequest')),
+			id = this.getCacheKey(opts),
 			cacheVal = instanceCache[id];
 
 		if (cacheVal != null) {
@@ -116,10 +116,6 @@ export default abstract class Provider extends ParamsProvider implements IProvid
 		this.cacheId = id;
 		this.async = new Async(this);
 		this.emitter = new EventEmitter({maxListeners: 1e3, newListener: false});
-
-		if (Object.isBoolean(opts.externalRequest)) {
-			this.setReadonlyParam('externalRequest', opts.externalRequest);
-		}
 
 		if (opts.socket || this.socketURL != null) {
 			this.connect().then(this.initSocketBehaviour.bind(this), stderr);
@@ -297,7 +293,6 @@ export default abstract class Provider extends ParamsProvider implements IProvid
 			method = this.method() ?? this.getMethod;
 
 		const mergedOpts = this.getRequestOptions('get', {
-			externalRequest: this.externalRequest,
 			...opts,
 			[queryMethods[method] != null ? 'query' : 'body']: query,
 			method