+++ src/core/request/index.ts
@@ -13,7 +13,6 @@
 
 import log from 'core/log';
 import Then from 'core/then';
-import { isOnline } from 'core/net';
 
 import Response from 'core/request/response';
 import RequestError from 'core/request/error';
@@ -163,13 +162,12 @@ function request<D = unknown>(
 			});
 
 			await new Promise(setImmediate);
+
 			ctx.parent = parent;
 
 			if (Object.isPromise(ctx.cache)) {
 				await Then.resolve(ctx.cache, parent);
 			}
-
-			await checkOnline();
 
 			const
 				tasks = <Array<CanPromise<unknown>>>[];
@@ -292,13 +290,8 @@ function request<D = unknown>(
 					decoders: ctx.decoders
 				};
 
-				const createReq = () => {
-					if (!ctx.isOnline && !requestParams.externalRequest) {
-						return Then.reject(new RequestError('offline', errDetails));
-					}
-
-					return requestParams.engine(reqOpts);
-				};
+				const
+					createReq = () => requestParams.engine(reqOpts);
 
 				if (requestParams.retry != null) {
 					const retryParams: RetryOptions = Object.isNumber(requestParams.retry) ?
@@ -330,7 +323,6 @@ function request<D = unknown>(
 						};
 
 						try {
-							await checkOnline();
 							return await createReq().then(wrapSuccessResponse);
 
 						} catch (err) {
@@ -361,7 +353,6 @@ function request<D = unknown>(
 			res
 				.then((response) => log(`request:response:${path}`, response.data, {
 					cache,
-					externalRequest: requestParams.externalRequest,
 					request: requestParams
 				}))
 
@@ -393,20 +384,12 @@ function request<D = unknown>(
 				const
 					data = await response.decode();
 
-				if (requestParams.externalRequest && !ctx.isOnline && data == null) {
-					throw new RequestError('offline', details);
-				}
-
 				return {
 					data,
 					response,
 					ctx,
 					dropCache: ctx.dropCache.bind(ctx)
 				};
-			}
-
-			async function checkOnline(): Promise<void> {
-				ctx.isOnline = (await Then.resolve(isOnline(), parent)).status;
 			}
 		});